Feat: addd defensive redirect

+++ kit/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import (
     render, redirect, reverse, get_object_or_404)
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
 from .models import Product, Category
@@ -71,10 +72,15 @@ def product_detail(request, product_id):
     return render(request, 'kit/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ 
     Add a product to the shop
     """
+    if not request.user.is_superuser:
+        message.error(request, 'Sorry. This action requires club admin access')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,10 +100,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ 
     Edit an exisiting product
     """
+    if not request.user.is_superuser:
+        message.error(request, 'Sorry. This action requires club admin access')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES,  instance=product)
@@ -121,10 +132,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ 
     Delete an exisiting product
     """
+    if not request.user.is_superuser:
+        message.error(request, 'Sorry. This action requires club admin access')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product has been deleted!')
